More warning fixes & pointer truncation fixes

+++ include/JSystem/JAudio2/JASHeapCtrl.h
@@ -265,13 +265,13 @@ namespace JASKernel {
     void setupRootHeap(JKRSolidHeap*, u32);
     JKRHeap* getSystemHeap();
     JASMemChunkPool<1024, JASThreadingModel::ObjectLevelLockable>* getCommandHeap();
-    void setupAramHeap(u32, u32);
+    void setupAramHeap(uintptr_t, u32);
     JASHeap* getAramHeap();
     u32 getAramFreeSize();
     u32 getAramSize();
 
     extern JASHeap audioAramHeap;
-    extern u32 sAramBase;
+    extern uintptr_t sAramBase;
     extern JKRHeap* sSystemHeap;
     extern JASMemChunkPool<1024, JASThreadingModel::ObjectLevelLockable>* sCommandHeap;
 };
