+++ src/f_op/f_op_scene_mng.cpp
@@ -29,14 +29,14 @@ fpc_ProcID fopScnM_DeleteReq(scene_class* i_scene) {
     return fopScnRq_Request(1, i_scene, 0x7FFF, NULL, 0x7FFF, 0) != fpcM_ERROR_PROCESS_ID_e;
 }
 
-int fopScnM_CreateReq(s16 i_procName, s16 param_2, u16 param_3, u32 i_data) {
+int fopScnM_CreateReq(s16 i_procName, s16 param_2, u16 param_3, uintptr_t i_data) {
     printf("[DIAG] fopScnM_CreateReq: procName=%d fade=%d\n", i_procName, param_2); fflush(stdout);
     fpc_ProcID result = fopScnRq_Request(0, 0, i_procName, (void*)i_data, param_2, param_3);
     printf("[DIAG] fopScnM_CreateReq: result=%d (error=%d)\n", result, fpcM_ERROR_PROCESS_ID_e); fflush(stdout);
     return result != fpcM_ERROR_PROCESS_ID_e;
 }
 
-u32 fopScnM_ReRequest(s16 i_procName, u32 i_data) {
+u32 fopScnM_ReRequest(s16 i_procName, uintptr_t i_data) {
     if (l_scnRqID == fpcM_ERROR_PROCESS_ID_e) {
         return 0;
     }
